text_writer_svc/requirements.txt
flask==2.3.3
kafka-python==2.0.2
psycopg2-binary==2.9.7
pyyaml==6.0.1
python-dotenv==1.0.0
click==8.1.7
requests==2.31.0